requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
arq==0.26.0
    # via -r requirements.in
build==1.2.1
    # via pip-tools
click==8.1.7
    # via
    #   arq
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
greenlet==3.0.3
    # via playwright
hiredis==3.0.0
    # via redis
packaging==24.1
    # via build
pip-tools==7.4.1
    # via -r requirements.in
playwright==1.45.1
    # via -r requirements.in
pyaes==1.6.1
    # via telethon
pyasn1==0.6.0
    # via rsa
pyee==11.1.0
    # via playwright
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via -r requirements.in
redis[hiredis]==4.6.0
    # via arq
rsa==4.9
    # via telethon
telethon==1.36.0
    # via -r requirements.in
typing-extensions==4.12.2
    # via pyee
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
